Revert "Fixed notifications... at least I hope so"
... no, I broke 'em completely :\

This reverts commit 6e811e04e598782f58676480764836fa59e80754.

+++ src/game/client/components/hud.h
@@ -10,7 +10,7 @@ class CHud : public CComponent
 	struct CNotification
 	{
 		char m_aMsg[256];
-		int64 m_SpawnTime;
+		float m_SpawnTime;
 		float m_xOffset;
 		vec4 m_Color;
